refactor: Decompose index.js into component-based structure
This major refactoring breaks down the monolithic index.js file into a more modular, component-based architecture. The primary goal is to improve code organization, maintainability, and separation of concerns.

Key changes include:
- **Componentization:** UI rendering logic has been extracted from `index.js` into dedicated component files under `src/components/`. This includes `Sidebar`, `MainChat`, `SettingsModal`, etc.
- **UI Rendering:** A new `src/ui.js` file now orchestrates the rendering of these components.
- **Event Handling:** Event listener logic has been moved into `src/handlers/` to decouple it from the main application class.
- **Utilities:** Helper functions for API calls (`api.js`), storage (`storage.js`), and general utilities (`utils.js`) have been separated into their own modules.

In addition to the refactoring, this commit also includes the following bug fixes that were identified during the process:

- **fix(render):** Corrected an issue where the UI would not render on the initial load due to the old and new states being identical. The `render` logic now ensures a full render on the first launch.
- **fix(init):** Removed a call to a non-existent `addKeyboardListeners` function that was causing a fatal error on startup.
